add fuzzy filtering to translation table
