+++ apps/web-main/src/components/table/FuzzyGlobalFilter.tsx
@@ -1,5 +1,7 @@
 import { Input } from '@design-system';
-import { RowData, Table } from '@tanstack/react-table';
+import { removeDiacritics } from '@lib-utils';
+import { rankItem } from '@tanstack/match-sorter-utils';
+import { FilterMeta, Row, RowData, Table } from '@tanstack/react-table';
 
 export const FuzzyGlobalFilter = <T extends RowData>({
   table,
@@ -16,4 +18,20 @@ export const FuzzyGlobalFilter = <T extends RowData>({
       className="max-w-sm"
     />
   );
+};
+
+export const fuzzyFilterFn = <T extends RowData>(
+  row: Row<T>,
+  columnId: string,
+  value: string,
+  addMeta: (meta: FilterMeta) => void,
+) => {
+  const rawFilter = removeDiacritics(value);
+  const rowValue = row.getValue(columnId);
+  const itemRank = rankItem(rowValue, rawFilter, {
+    keepDiacritics: false,
+  });
+  addMeta({ itemRank });
+  if (itemRank.passed) console.log(itemRank);
+  return itemRank.passed && itemRank.rank > 2;
 };
